Cast hc_pre NPU op output back to input dtype
torch.ops.custom.npu_hc_pre emits y in fp32 (sinkhorn iterates in fp32),
but the next op in the V4 forward chain is input_layernorm which calls
aclnnRmsNorm with the layer's bf16 gamma. ACL has no
x=fp32 + gamma=bf16 RmsNorm overload (only fp32/fp32, bf16/fp32,
bf16/bf16, fp16/fp32, fp16/fp16), so the call fails with error 161002.

Mirror the torch fallback's 'y.to(dtype)' cast and the bug clears.

+++ python/sglang/srt/models/deepseek_v4.py
@@ -702,7 +702,12 @@ class DeepseekV4DecoderLayer(nn.Module):
                 norm_eps=self.rms_norm_eps,
                 hc_eps=self.hc_eps,
             )
-            return y, post, comb
+            # The fused kernel emits y in fp32 (the sinkhorn iterates in
+            # fp32). The downstream input_layernorm calls aclnnRmsNorm with
+            # bf16 gamma; ACL has no x=fp32 / gamma=bf16 overload, so we
+            # must cast back to the original dtype here. Mirrors the torch
+            # path's `y.to(dtype)` below.
+            return y.to(dtype), post, comb
 
         if envs.SGLANG_OPT_USE_TILELANG_MHC_PRE.get():
             from sglang.srt.layers.mhc import mhc_pre
